pyproject.toml
[tool.poetry]
name = "concrexit"
version = "38.0"
description = "New new Thalia website, now with extra Django."
authors = ["Thalia Technicie <www@thalia.nu>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
django-localflavor = "3.1"
freezegun = "1.1.0"
bleach = "4.1.0"
django-tinymce4-lite = "1.8.0"
djangorestframework = "3.12.4"
django-ical = "1.8.0"
django-libsass = "0.8.0"
python-magic = "0.4.24"
Django = "3.2.8"
Pillow = "8.4.0"
django-compressor = "2.4.1"
psycopg2-binary = "2.9.1"
bcrypt = "3.2.0"
argon2_cffi = "21.1.0"
uWSGI = "2.0.20"
django-bootstrap4 = "3.0.1"
firebase-admin = "5.0.3"
sentry-sdk = "1.4.3"
django-sendfile2 = "0.6.0"
django-queryable-properties = "^1.6.1"

# docs requirements
recommonmark = { version = "0.7.1", optional = true }
sphinx = { version = "4.2.0", optional = true }
google-api-python-client = "2.26.1"
django-oauth-toolkit = "^1.3.2"
django-cors-headers = "^3.5.0"
PyYAML = "^5.3.1"
django-debug-toolbar = "^3.1.1"
pydocstyle = "^6.0.0"
django-admin-autocomplete-filter = "^0.7"
django-filter = "^21.1"

[tool.poetry.extras]
docs = ["recommonmark", "sphinx"]

[tool.poetry.dev-dependencies]
django-template-check = "0.4.0"
factory_boy = "3.2.0"
pydenticon = "0.3.1"
Faker = "9.3.1"
coverage = {extras = ["toml"], version = "^5.0.3"}
black = "19.10b0"
pylint = "^2.6.0"
pylint-django = "^2.3.0"

[tool.coverage.run]
branch = true
source = ["website"]
omit = ["website/manage.py", "website/thaliawebsite/wsgi.py", "website/thaliawebsite/settings.py", "website/*/migrations/*", "website/*/test*"]

[tool.coverage.report]
precision = 2
skip_covered = true
show_missing = true
fail_under = 65.0

[tool.black]
exclude = '(/(\.eggs|\.git|\.tox)/|migrations)'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
